requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
email-validator==2.1.0
pydantic[email]==2.5.0
aiofiles==24.1.0
matplotlib==3.7.2
seaborn==0.12.2